# docs/swagger.yaml
basePath: /api/v1
definitions:
  AddMemberRequest:
    properties:
      is_admin:
        example: false
        type: boolean
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - user_id
    type: object
  CreateOrganizationRequest:
    properties:
      address:
        example: 254 St, Bangkok, TH
        type: string
      description:
        example: Higher education institution
        type: string
      image_urls:
        example:
        - https://example.com/example-1.jpg
        - https://example.com/example-2.jpg
        items:
          type: string
        type: array
      lat:
        example: 13.7388
        type: number
      lng:
        example: 100.5322
        type: number
      name:
        example: Acme Corp
        type: string
    required:
    - lat
    - lng
    - name
    type: object
...
  ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    required:
    - error
    type: object
  GetOrganizationsByIDsRequest:
    properties:
      ids:
        example:
        - 550e8400-e29b-41d4-a716-446655440000
        - 550e8400-e29b-41d4-a716-446655440001
        items:
          type: string
        type: array
    required:
    - ids
    type: object
  OrganizationCoord:
    properties:
      id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      lat:
        example: 13.7388
        type: number
      lng:
        example: 100.5322
        type: number
    required:
    - id
    - lat
    - lng
    type: object
  OrganizationResponse:
    properties:
      address:
        example: 254 St, Bangkok, TH
        type: string
      created_at:
        example: "2026-01-01T12:00:00.00000+07:00"
        type: string
      description:
        example: Higher education institution
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      image_urls:
        example:
        - https://example.com/example-1.jpg
        - https://example.com/example-2.jpg
        items:
          type: string
        type: array
      lat:
        example: 13.7888
        type: number
      lng:
        example: 100.5322
        type: number
      name:
        example: Acme Corp
        type: string
      updated_at:
        example: "2026-01-01T12:00:00.00000+07:00"
        type: string
    required:
    - created_at
    - id
    - image_urls
    - lat
    - lng
    - name
    - updated_at
    type: object
  SetAdminRequest:
    properties:
      is_admin:
        example: true
        type: boolean
    required:
    - is_admin
    type: object
  UpdateOrganizationRequest:
    properties:
      address:
        example: 254 St, Bangkok, TH
        type: string
      description:
        example: Higher education institution
        type: string
      image_urls:
        example:
        - https://example.com/example-1.jpg
        - https://example.com/example-2.jpg
        items:
          type: string
        type: array
      lat:
        example: 13.7388
        type: number
      lng:
        example: 100.5322
        type: number
      name:
        example: Acme Corp
        type: string
    type: object
...
  UserMembershipResponse:
    properties:
      is_admin:
        example: false
        type: boolean
      organization_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      organization_name:
        example: Acme Corp
        type: string
    type: object
  UserResponse:
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      created_at:
        example: "2026-01-01T12:00:00.00000+07:00"
        type: string
      description:
        example: Senior researcher specializing in quantum computing
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: John Doe
        type: string
      organizations:
        items:
          $ref: '#/definitions/UserMembershipResponse'
        type: array
      phone_number:
        example: "+1234567890"
        type: string
      research_categories:
        example:
        - QuantumComputing
        - Qiskit
        - Cryogenics
        items:
          type: string
        type: array
      social_media:
        example: '@john on Twitter, linkedin.com/in/john'
        type: string
      updated_at:
        example: "2026-01-01T12:00:00.00000+07:00"
        type: string
    required:
    - created_at
    - email
    - id
    - name
    - research_categories
    - updated_at
    type: object
info:
  contact: {}
  description: A simple REST API for managing users and organizations
  title: Customer API
  version: "1.0"
paths:
  /organizations:
    get:
      consumes:
      - application/json
      description: Get a list of all organizations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/OrganizationResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get all organizations
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: Create a new organization with name, location, and optional details
      parameters:
      - description: Organization to create
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/OrganizationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create a new organization
      tags:
      - organizations
  /organizations/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete an organization by ID
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Delete an organization
      tags:
      - organizations
    get:
      consumes:
      - application/json
      description: Get a single organization by their ID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/OrganizationResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get an organization by ID
      tags:
      - organizations
    patch:
      consumes:
      - application/json
      description: Update an existing organization by ID (partial updates supported)
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/UpdateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/OrganizationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update an organization
      tags:
      - organizations
  /organizations/{id}/members:
    get:
      consumes:
      - application/json
      description: Get all users that belong to an organization
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/UserResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get members of an organization
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: Add an existing user to an organization with an optional admin
        role
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Member to add
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/AddMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Add a member to an organization
      tags:
      - organizations
  /organizations/{id}/members/{user_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Remove a member from an organization
      tags:
      - organizations
    patch:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Admin flag
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/SetAdminRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update a member's admin role
      tags:
      - organizations
  /organizations/batch:
    post:
      consumes:
      - application/json
      description: Get multiple organizations by their UUIDs in a single request
      parameters:
      - description: List of organization IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/GetOrganizationsByIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/OrganizationResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get organizations by multiple IDs (batch)
      tags:
      - organizations
  /organizations/coordinates:
    get:
      consumes:
      - application/json
      description: Get ID and coordinates of all organizations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/OrganizationCoord'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get all organization coordinates
      tags:
      - organizations
  /organizations/search:
    get:
      consumes:
      - application/json
      description: Search organizations by name using ILIKE query
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: 'Maximum number of results to return (default: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/OrganizationResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Search organizations
      tags:
      - organizations
...
swagger: "2.0"
...